[fix] use self.pk instead of self.id
some models don't use an id field. If a model is based off of
AbstractCA/Cert and specifies a primary_key field that isn't
the default `id`, Attribute errors will occur in the save method.

Swap `self.id` for `self.pk` will still point to `id` if it's not
specified, but will allow for users to use a different PK field.

https://docs.djangoproject.com/en/2.2/ref/models/instances/#django.db.models.Model.pk