[pki] Add possibility of renewing certificates #83
Added a new function in the base class which
renews the certificate, private key and
validity end date by calling the generate
function and updating the validity end date.
Also added actions in CA and cert admins to
make use of this newly added function.
Also, a confirmation page was added to the
renew actions which Fixes #85

Fixes #83